+++ src/services/PrivateKeyService.js
@@ -15,6 +15,8 @@ class EthereumAddressError extends BaseError {}
 class InvalidPrivateKeyError extends BaseError {}
 class InvalidPrivateKeyFile extends BaseError {}
 class InvalidMnemonicError extends BaseError {}
+class InvalidKeystoreFile extends BaseError {}
+class InvalidKeystoreParams extends BaseError {}
 
 const HD_PATH = "m/44'/60'/0/0"
 const encoding = 'utf8'
@@ -45,7 +47,7 @@ class PrivateKeyService {
   }
 
   /**
-   * It gets the private key reading it from file.
+   * It gets a palin private key reading it from file.
    *
    * @param {String} privateKeyPath  The path of private key file.
    *
@@ -68,6 +70,49 @@ class PrivateKeyService {
   }
 
   /**
+   * It gets the keystore object from a keystore file.
+   *
+   * @param {String} keystoreFilePath  The path of the keystore.
+   *
+   * @throws {InvalidKeystoreFile} If does not exist the file.
+   */
+  async getKeystoreAsync(keystoreFilePath) {
+    try {
+      const extractedData = await readFileAsync(keystoreFilePath, encoding)
+      // Regex for both windows and unix system.
+      const keyStoreObject = extractedData.split(/[\r\n]+/).shift()
+      this.log.debug({ keystoreFilePath, fn: 'getKeyStoreAsync' },
+        'Get private keystore from file.')
+      return JSON.parse(keyStoreObject)
+    } catch (err) {
+      this.log.error({ keystoreFilePath, fn: 'getKeyStoreAsync' },
+        'Error getting keystore from file.')
+      throw new InvalidKeystoreFile(err)
+    }
+  }
+
+  /**
+   * It gets the private key from a keystore.
+   *
+   * @param {String} password        The password to decrypt the keystore.
+   * @param {String} keystore        The keystore.
+   *
+   * @throws {InvalidKeystoreParams} If something goes wrong during the recovering.
+   */
+  async getPrivateKeyFromKeystore(password, keystore) {
+    try {
+      const bufferPrivateKey = keythereum.recover(password, keystore)
+      const privateKey = bufferPrivateKey.toString('hex')
+      this.log.info({ fn: 'getPrivateKeyFromKeystore', privateKey }, 'Get privatekey from keystore using the password.')
+      return privateKey
+    } catch (err) {
+      this.log.error({ keystore, fn: 'getPrivateKeyFromKeystore' },
+        'Error recovering keystore using the password.')
+      throw new InvalidKeystoreParams(err)
+    }
+  }
+
+  /**
    * It gets the address from private key.
    *
    * @param {String} privateKey  The  private key.
@@ -76,7 +121,8 @@ class PrivateKeyService {
    */
   getAddressFromPrivateKey(privateKey) {
     try {
-      const privateKeyBuffered = ethereumUtil.toBuffer(privateKey)
+      const privateKeyWithPrefix = ethereumUtil.addHexPrefix(privateKey)
+      const privateKeyBuffered = ethereumUtil.toBuffer(privateKeyWithPrefix)
       const addressBufferedFromPrivateKey = ethereumUtil.privateToAddress(privateKeyBuffered)
       const personalWalletAddress = ethereumUtil.bufferToHex(addressBufferedFromPrivateKey)
       this.log.debug({ fn: 'getAddressFromPrivateKey' },
@@ -141,4 +187,11 @@ class PrivateKeyService {
   }
 }
 
-module.exports = { PrivateKeyService, InvalidPrivateKeyError, InvalidPrivateKeyFile, InvalidMnemonicError }
+module.exports = {
+  InvalidKeystoreFile,
+  InvalidKeystoreParams,
+  InvalidMnemonicError,
+  InvalidPrivateKeyError,
+  InvalidPrivateKeyFile,
+  PrivateKeyService,
+}