+++ src/validators/PrivateKeyValidator.js
@@ -7,8 +7,14 @@ const privateKeySchema = customJoiValidator.object().keys({
   privateKey: customJoiValidator.privateKey().ethereumValidPrivateKey().required(),
 })
 
+// TODO: to be removed as soon as the migration to keystore is finished
 const privateKeyFilePathSchema = customJoiValidator.object().keys({
   privateKeyFilePath: customJoiValidator.path()
+    .existFile().required(),
+})
+
+const keystoreFilePathSchema = customJoiValidator.object().keys({
+  keystoreFilePath: customJoiValidator.path()
     .existFile().required(),
 })
 
@@ -17,9 +23,15 @@ class PrivateKeyValidator extends BaseValidator {
     return this.constructor.validate(privateKey, privateKeySchema)
   }
 
+  // TODO: to be removed as soon as the migration to keystore is finished
   validateFilePath(privateKeyFilePath) {
     this.constructor.validate(privateKeyFilePath, privateKeyFilePathSchema)
     return privateKeyFilePath
+  }
+
+  validateKeystoreFilePath(keystoreFilePath) {
+    this.constructor.validate(keystoreFilePath, keystoreFilePathSchema)
+    return keystoreFilePath
   }
 
   validateMnemonic(mnemonic) {