+++ src/validators/commands/ethereum-wallet/KeyStoreGenerateCommandValidator.js
@@ -4,8 +4,8 @@ const customJoiValidator = require('../../../utils/customJoiValidator')
 const keyStoreGenerateCommandSchema = customJoiValidator.object()
   .keys({
     hdPath: customJoiValidator.string().required(),
-    keyStoreFilePath: customJoiValidator.string().required(),
-    keyStorePassword: customJoiValidator.string().required(),
+    keystoreFilePath: customJoiValidator.string().required(),
+    keystorePassword: customJoiValidator.string().required(),
     mnemonic: customJoiValidator.mnemonic().ethereumValidMnemonic().required(),
   })